Create /helpers directory
