+++ scripts/helpers/categories.json
